feat: file/attachment context menu (#479)
* feat: file context menu

Works on Images, Videos and Audio files. I decided to leave out text
files is because preferably we would use the browser's context menu for
text selection.

Signed-off-by: Amy <amy+git@amogus.cloud>

* doc: explain why SaveFile opens the originalUrl instead of initiating a
download

Signed-off-by: Amy <amy+git@amogus.cloud>

* chore: add icons to context menu entries

Signed-off-by: Amy <amy+git@amogus.cloud>

* chore: move attachment options to MessageContextMenu

* chore: provide message prop on attachment

So we can display the message context menu + the attachment options

* fix: forgot to wrap save file option in anchor

Signed-off-by: Amy <amy+git@amogus.cloud>

* fix: make message optional in context menu

The reason why is because Attachment is inside of TextEmbed, and passing
the parent message through two components doesn't seem too efficient and
might make merge conflicts. So we only show message options when message
is passed to the context menu.

Signed-off-by: Amy <amy+git@amogus.cloud>

---------

Signed-off-by: Amy <amy+git@amogus.cloud>
Co-authored-by: izzy <me@insrt.uk>